+++ packages/sit-onyx/src/components/OnyxAppLayout/types.ts
@@ -1,0 +1,4 @@
+export type OnyxAppLayoutProps = {
+  /** Whether the nav bar will stick to the left or to the top of the app */
+  navBarAlignment?: "top" | "left";
+};