+++ apps/alpha-test-app/CHANGELOG.md
@@ -1,5 +1,19 @@
 # alpha-test-app
 
+## 0.0.1-alpha.24
+
+### Patch Changes
+
+- Updated dependencies [827a893]
+  - sit-onyx@1.0.0-alpha.27
+
+## 0.0.1-alpha.23
+
+### Patch Changes
+
+- Updated dependencies [2bc861c]
+  - sit-onyx@1.0.0-alpha.26
+
 ## 0.0.1-alpha.22
 
 ### Patch Changes